+++ R/auth.R
@@ -22,6 +22,10 @@
 #' @export
 armadillo.login <- function(armadillo, minio, duration = 900) { # nolint
   token <- armadillo.get_token(armadillo)
+  options(
+    MolgenisArmadillo.auth.token = token,
+    MolgenisArmadillo.armadillo.endpoint = armadillo
+  )
   armadillo.assume_role_with_web_identity(token, minio, duration)
   invisible(token)
 }