+++ R/utils.R
@@ -63,3 +63,22 @@
 .use_https <- function() {
   getOption("MolgenisArmadillo.s3.use_https", TRUE)
 }
+
+#' Handle generic request errors
+#'
+#' @param response HTTR response
+#'
+#' @importFrom httr content
+#'
+#' @noRd
+.handle_request_error <- function(response) {
+  if (response$status_code == 401) {
+    stop("Unauthorized", call. = FALSE)
+  } else if (response$status_code == 500) {
+    stop(
+      paste0(
+        "Internal server error: ",
+        httr::content(response, as = "text", encoding = "UTF-8")),
+        call. = FALSE)
+  }
+}