+++ NEWS.md
@@ -1,3 +1,20 @@
+# MolgenisArmadillo (development version)
+
+# MolgenisArmadillo 1.1.3
+
+# MolgenisArmadillo 1.1.2
+
+# MolgenisArmadillo 1.1.1
+
+# MolgenisArmadillo 1.1.0
+## Features
+### Developer tools
+- Install packages
+- Whitelist packages
+## Chores
+- Make Mariska Slofstra maintainer
+
+
 # MolgenisArmadillo 1.0.3
 
 # MolgenisArmadillo 1.0.2